wrong datatype in lfoExtended::setToStep()

+++ lfoExtended.cpp
@@ -48,7 +48,7 @@ void lfoExtended::setThreshold(uint8_t thres,LFOThresholdType type) {
 }
 
 
-void lfoExtended::setToStep(uint8_t step, uint8_t timestamp) {
+void lfoExtended::setToStep(uint8_t step, uint16_t timestamp) {
 	timestampOffset = timestamp - ((step*bastlCyclesPerPeriod)>>8);
 	#ifdef TESTING
 	//printf("Offset: %u\n",timestampOffset);
